# Improve performance/decrease allocations in hWnd proxy for WindowsMenu