--- a/src/Microsoft.DotNet.Wpf/src/UIAutomation/UIAutomationClientSideProviders/MS/Internal/AutomationProxies/Misc.cs
+++ b/src/Microsoft.DotNet.Wpf/src/UIAutomation/UIAutomationClientSideProviders/MS/Internal/AutomationProxies/Misc.cs
@@ -11,8 +11,8 @@
 using Microsoft.Win32.SafeHandles;
 using MS.Win32;
 using System;
+using System.Globalization;
 using System.ComponentModel;
-using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Text;
 using System.Windows.Automation;
@@ -1660,7 +1660,7 @@ namespace MS.Internal.AutomationProxies
 
             // Update '&' index and slice it out
             ampersandIndex -= (value.Length - trimmed.Length);
-            return string.Create(null, stackalloc char[128], $"{trimmed.Slice(0, ampersandIndex)}{trimmed.Slice(ampersandIndex + 1)}");
+            return string.Create(CultureInfo.InvariantCulture, stackalloc char[128], $"{trimmed.Slice(0, ampersandIndex)}{trimmed.Slice(ampersandIndex + 1)}");
         }
 
         internal static void ThrowWin32ExceptionsIfError(int errorCode)