## Description

- Replaces string concatenations with string interpolation throughout the whole proxy class.
- Also modifies `StripMnemonic` helper function from the `Misc` class used by other proxies with accelerators.
- Retrievals of `LocalizedName` and `AcceleratorKey` were highly optimized.

### AccelatorKeyCtrl benchmark
| Method   | findPattern | Mean [ns] | Error [ns] | StdDev [ns] | Gen0   | Code Size | Allocated [B] |
|--------- |------------ |----------:|-----------:|------------:|-------:|--------------:|--------------:|
| Original | Win+A       | 205.74 ns |   2.361 ns |    2.093 ns | 0.0277 |       4,753 B |         464 B |
| PR__EDIT | Win+A       |  97.97 ns |   0.519 ns |    0.485 ns | 0.0024 |       1,105 B |          40 B |

<details><summary>Benchmark code</summary>

```csharp
[Benchmark]
[Arguments("Win+A")]
public string Original(string findPattern)
{
    int pos;
    string keyCtrl = "Ctrl";
    string keyControl = "Control";
    string keyAlt = "Alt";
    string keyShift = "Shift";
    string keyWin = "Win";

    string menuText = findPattern.ToLower(CultureInfo.InvariantCulture);
    string accelerator;

    if ((accelerator = AccelatorKeyCtrl(keyCtrl.ToLower(CultureInfo.InvariantCulture), keyCtrl + " + ", menuText, findPattern, out pos)) != null ||
        (accelerator = AccelatorKeyCtrl(keyControl.ToLower(CultureInfo.InvariantCulture), keyCtrl + " + ", menuText, findPattern, out pos)) != null ||
        (accelerator = AccelatorKeyCtrl(keyAlt.ToLower(CultureInfo.InvariantCulture), keyAlt + " + ", menuText, findPattern, out pos)) != null ||
        (accelerator = AccelatorKeyCtrl(keyShift.ToLower(CultureInfo.InvariantCulture), keyShift + " + ", menuText, findPattern, out pos)) != null ||
        (accelerator = AccelatorKeyCtrl(keyWin.ToLower(CultureInfo.InvariantCulture), keyWin + " + ", menuText, findPattern, out pos)) != null)
    {
        return accelerator;
    }

    return null;
}

[Benchmark]
[Arguments("Win+A")]
public string PR__EDIT(string findPattern)
{
    string accelerator;
    if ((accelerator = AccelatorKeyCtrlNew("Ctrl", findPattern, out _)) != null ||
      (accelerator = AccelatorKeyCtrlNew("Control", findPattern, out _)) != null ||
      (accelerator = AccelatorKeyCtrlNew("Alt", findPattern, out _)) != null ||
      (accelerator = AccelatorKeyCtrlNew("Shift", findPattern, out _)) != null ||
      (accelerator = AccelatorKeyCtrlNew("Win", findPattern, out _)) != null)
    {
        return accelerator;
    }

    return null;
}
```

</details>

### AccelatorFxx benchmark
| Method   | findPattern       | Mean [ns] | Error [ns] | StdDev [ns] | Gen0   | Code Size | Allocated [B] |
|--------- |------------------ |----------:|-----------:|------------:|-------:|--------------:|--------------:|
| Original | xxxxxxAAAxxxxxF11 |  53.52 ns |   1.108 ns |    1.758 ns | 0.0105 |       2,389 B |         176 B |
| PR__EDIT | xxxxxxAAAxxxxxF11 |  31.61 ns |   0.598 ns |    0.560 ns | 0.0052 |       1,338 B |          88 B |

<details><summary>Benchmark code</summary>

```csharp
[Benchmark]
[Arguments("xxxxxxAAAxxxxxF11")]
public string Original(string findPattern)
{
    // Try to look for a Fxx
    int pos;
    string menuText = findPattern.ToLower(CultureInfo.InvariantCulture);
    string accelerator = AccelatorFxx(menuText);
    if (!string.IsNullOrEmpty(accelerator))
    {
        pos = menuText.LastIndexOf(accelerator, StringComparison.OrdinalIgnoreCase);
        if (pos >= 0)
        {
            return findPattern.Substring(0, pos);
        }
        else
        {
            // Wrong logic, we should be able to find the Fxx combination we just built
            System.Diagnostics.Debug.Assert(false, "Cannot find back the accelerator in the menu!");
            return findPattern;
        }
    }

    return null;
}

[Benchmark]
[Arguments("xxxxxxAAAxxxxxF11")]
public string PR__EDIT(string findPattern)
{
    // Try to look for a Fxx
    int pos;
    string accelerator = AccelatorFxxNew(findPattern, out pos);
    if (!string.IsNullOrEmpty(accelerator))
    {
        return findPattern.Substring(0, pos);
    }

    return null;
}
```

</details>

### StripMnemonic benchmark
| Method   | findPattern | Mean [ns] | Error [ns] | StdDev [ns] | Gen0   | Allocated [B] |
|--------- |------------ |----------:|-----------:|------------:|-------:|--------------:|
| Original | wow F11     | 26.071 ns |  0.5511 ns |   0.8248 ns | 0.0067 |                 112 B |
| PR__EDIT | wow F11     |  2.992 ns |  0.0743 ns |   0.0695 ns |      - |                  - |
| Original |  \x20wow F11    | 25.000 ns |  0.4421 ns |   0.4135 ns | 0.0067 |              112 B |
| PR__EDIT |  \x20wow F11    |  9.294 ns |  0.2081 ns |   0.1737 ns | 0.0024 |                 40 B |
| Original | &wow F11    |  25.59 ns |   0.494 ns |    0.643 ns | 0.0067 |                112 B |
| PR__EDIT | &wow F11    |  16.69 ns |   0.259 ns |    0.242 ns | 0.0024 |                40 B |
| Original | \x20\x20\x20\x20\x20&wow F11 |  27.03 ns |   0.533 ns |    0.473 ns | 0.0076 |                 128 B |
| PR__EDIT | \x20\x20\x20\x20\x20&wow F11 |  17.36 ns |   0.328 ns |    0.291 ns | 0.0024 |               40 B |

### Find from list of pre-defined accelerators
| Method   | findPattern           | Mean [ns] | Error [ns] | StdDev [ns] | Gen0   | Code Size | Allocated [B] |
|--------- |---------------------- |----------:|-----------:|------------:|-------:|--------------:|--------------:|
| Original |  wowo(...) \tTAB [45] |  828.5 ns |    9.83 ns |     9.19 ns | 0.0067 |         721 B |         112 B |
| PR__EDIT |  wowo(...) \tTAB [45] |  537.9 ns |    4.06 ns |     3.80 ns |      - |         485 B |             - |

<details><summary>Benchmark code</summary>

```csharp
private static readonly string[] s_keywordsAccelerators = ["Home", "End", "Delete", "Page Up", "Page Down", "Escape", "Insert", "Print Screen", "Tab"];

[Benchmark]
[Arguments(" wowowwowowowowowowowowwowowowowowowowow \tTAB")]
public string Original(string findPattern)
{
    int pos;
    string menuText = findPattern.ToLower(CultureInfo.InvariantCulture);
    for (int i = 0; i < s_keywordsAccelerators.Length; i++)
    {
        pos = menuText.LastIndexOf(s_keywordsAccelerators[i], StringComparison.OrdinalIgnoreCase);
        if (pos > 0 && pos + s_keywordsAccelerators[i].Length == menuText.Length && (menuText[pos - 1] == '\a' || menuText[pos - 1] == '\t'))
        {
            return s_keywordsAccelerators[i];
        }
    }

    return null;
}

[Benchmark]
[Arguments(" wowowwowowowowowowowowwowowowowowowowow \tTAB")]
public string PR__EDIT(string findPattern)
{
    int pos;
    for (int i = 0; i < s_keywordsAccelerators.Length; i++)
    {
        pos = findPattern.LastIndexOf(s_keywordsAccelerators[i], StringComparison.InvariantCultureIgnoreCase);
        if (pos > 0 && pos + s_keywordsAccelerators[i].Length == findPattern.Length && (findPattern[pos - 1] == '\a' || findPattern[pos - 1] == '\t'))
        {
            return s_keywordsAccelerators[i];
        }
    }

    return null;
}
```

</details>

## Customer Impact

Increased performance/decreased allocation when using the proxy.

## Regression

No.

## Testing

Local build, CI, extensive local assert testing.

## Risk

Should be low, however reviewers should check I haven't changed the output meaning of each operation (culture handling).